use makeFunction instead of patch
