+++ apps/meteor/app/livechat/server/lib/Contacts.ts
@@ -18,6 +18,7 @@ import {
 	Subscriptions,
 	LivechatContacts,
 } from '@rocket.chat/models';
+import { makeFunction } from '@rocket.chat/patch-injection';
 import type { PaginatedResult, VisitorSearchChatsResult } from '@rocket.chat/rest-typings';
 import { check } from 'meteor/check';
 import { Meteor } from 'meteor/meteor';
@@ -55,6 +56,15 @@ type CreateContactParams = {
 	customFields?: Record<string, string | unknown>;
 	contactManager?: string;
 	channels?: ILivechatContactChannel[];
+};
+
+type VerifyContactChannelParams = {
+	contactId: string;
+	field: string;
+	value: string;
+	channelName: string;
+	visitorId: string;
+	roomId: string;
 };
 
 type UpdateContactParams = {
@@ -535,3 +545,9 @@ export async function validateContactManager(contactManagerUserId: string) {
 		throw new Error('error-contact-manager-not-found');
 	}
 }
+
+export const verifyContactChannel = makeFunction(async (_params: VerifyContactChannelParams): Promise<ILivechatContact | null> => null);
+
+export const mergeContacts = makeFunction(
+	async (_contactId: string, _channel: ILivechatContactChannel): Promise<ILivechatContact | null> => null,
+);
