--- apps/meteor/ee/app/contact-id-verification/server/index.ts
@@ -1,1 +1,0 @@
-import './services/omnichannelService';
