+++ apps/meteor/ee/server/patches/index.ts
@@ -1,3 +1,5 @@
 import './closeBusinessHour';
 import './getInstanceList';
 import './isDepartmentCreationAvailable';
+import './verifyContactChannel';
+import './mergeContacts';
